requirements.txt
alembic==1.16.1
arabic-reshaper==3.0.0
asn1crypto==1.5.1
bcrypt==4.0.1
blinker==1.9.0
Brotli==1.1.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
cryptography==46.0.2
cssselect2==0.8.0
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-Principal==0.4.0
Flask-Security-Too==5.3.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.60.1
freetype-py==2.5.1
greenlet==3.2.2
gunicorn==21.2.0
html5lib==1.1
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
oscrypto==1.3.0
packaging==25.0
passlib==1.7.4
pillow==11.3.0
psycopg2-binary==2.9.10
pycairo==1.28.0
pycparser==2.23
pydyf==0.11.0
pyHanko==0.31.0
pyhanko-certvalidator==0.29.0
PyJWT==2.10.1
pypdf==6.1.1
pyphen==0.17.2
python-bidi==0.6.6
python-dotenv==1.1.0
python-slugify==8.0.4
PyYAML==6.0.3
reportlab==4.4.3
requests==2.32.5
rlPyCairo==0.4.0
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.41
svglib==1.6.0
text-unidecode==1.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tomli==2.2.1
typing_extensions==4.13.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
weasyprint==66.0
webencodings==0.5.1
Werkzeug==3.1.2
WTForms==3.2.1
zopfli==0.2.3.post1
